+++ karma.conf.js
@@ -149,6 +149,11 @@ module.exports = function(config) {
                     if (ch > -1) {
                         availableBrowsers[ch] = 'Chromium_dev';
                     }
+                    // ignore IE
+                    let ie = availableBrowsers.indexOf('IE');
+                    if (ie > -1) {
+                        availableBrowsers.splice(ie, 1);
+                    }
                     return result;
                 }
             }
@@ -165,7 +170,7 @@ module.exports = function(config) {
         webpack: webpackConfig,
         customLaunchers: {
             Chrome_dev: {
-                base: 'ChromeHeadless',
+                base: 'Chrome',
                 flags: chromeFlags
             },
             Chromium_dev: {