+++ src/js/plugins/opus-recorder-plugin.js
@@ -11,11 +11,17 @@ const RecordEngine = videojs.getComponent('RecordEngine');
  * Audio is encoded using libopus.
  *
  * @class
- * @augments videojs.RecordEngine
+ * @augments RecordEngine
  */
 class OpusRecorderEngine extends RecordEngine {
     /**
      * Setup recording engine.
+     *
+     * @param {LocalMediaStream} stream - Media stream to record.
+     * @param {Object} mediaType - Object describing the media type of this
+     *     engine.
+     * @param {Boolean} debug - Indicating whether or not debug messages should
+     *     be printed in the console.
      */
     setup(stream, mediaType, debug) {
         this.inputStream = stream;
@@ -76,6 +82,7 @@ class OpusRecorderEngine extends RecordEngine {
 
     /**
      * @private
+     * @param {Object} data - Audio data returned by opus-recorder.
      */
     onRecordingAvailable(data) {
         // Opus format stored in an Ogg container