+++ test/test-helpers.js
@@ -96,13 +96,13 @@ const TestHelpers = {
             }
         };
         opts = mergeOptions(opts, newOptions);
-        var tag = TestHelpers.makeTag('audio', 'audioOnly');
+        let tag = TestHelpers.makeTag('audio', 'audioOnly');
         return this.makePlayer(tag, opts);
     },
 
     makeAudioOnlyPluginPlayer(pluginName) {
-        var tag = TestHelpers.makeTag('audio', 'audioOnly');
-        var recordPluginOptions = {
+        let tag = TestHelpers.makeTag('audio', 'audioOnly');
+        let recordPluginOptions = {
             audio: true,
             video: false,
             maxLength: 5,
@@ -174,7 +174,7 @@ const TestHelpers = {
             }
         };
         opts = mergeOptions(opts, newOptions);
-        var tag = TestHelpers.makeTag('video', 'audioVideo');
+        let tag = TestHelpers.makeTag('video', 'audioVideo');
         return this.makePlayer(tag, opts);
     },
 
@@ -201,7 +201,7 @@ const TestHelpers = {
     },
 
     makeImageOnlyPlayer() {
-        var tag = TestHelpers.makeTag('video', 'imageOnly');
+        let tag = TestHelpers.makeTag('video', 'imageOnly');
         return this.makePlayer(tag, {
             controls: true,
             autoplay: false,
@@ -225,7 +225,7 @@ const TestHelpers = {
     makeScreenOnlyPlayer(newOptions) {
         // use polyfill in Firefox for now, see:
         // https://blog.mozilla.org/webrtc/getdisplaymedia-now-available-in-adapter-js/
-        if (browserDetails.browser == 'firefox') {
+        if (browserDetails.browser === 'firefox') {
             browserShim.shimGetDisplayMedia(window, 'screen');
         }
         let opts = {
@@ -251,7 +251,7 @@ const TestHelpers = {
     },
 
     makeAnimatedPlayer() {
-        var tag = TestHelpers.makeTag('video', 'animationOnly');
+        let tag = TestHelpers.makeTag('video', 'animationOnly');
         return this.makePlayer(tag, {
             controls: true,
             autoplay: false,