+++ test/utils/browser-shim.spec.js
@@ -6,9 +6,5 @@ import setSrcObject from '../../src/js/utils/browser-shim.js';
 
 /** @test {browser-shim} */
 describe('utils.setSrcObject', () => {
-
-    /** @test {setSrcObject} */
-    it('works', () => {
-        let obj = setSrcObject({}, {}, false);
-    });
+    // todo
 });